Try to fix serialization error

+++ evolution/src/main/java/com/evolving/nglm/evolution/ThirdPartyManager.java
@@ -4963,13 +4963,13 @@ public class ThirdPartyManager
             // 2. subscriberID == null and event is an autoprovision event, so in this case, we provide in the constructor, the alternateID coming from the REST request
             if(subscriberID != null)
               {
-                eev = constructor.newInstance(new Object[]{subscriberID, SystemTime.getCurrentTime(), eventBody, tenantID });
+                eev = constructor.newInstance(new Object[]{subscriberID, SystemTime.getCurrentTime(), eventBody, (short) tenantID });
                 eev.forceDeliveryPriority(DELIVERY_REQUEST_PRIORITY);
               }
             else if(autoProvisionEvent != null)
               {
                 // means this is an autoprovision event, case 2...
-                eev = constructor.newInstance(new Object[]{subscriberParameter.getSecondElement(), SystemTime.getCurrentTime(), eventBody, tenantID });
+                eev = constructor.newInstance(new Object[]{subscriberParameter.getSecondElement(), SystemTime.getCurrentTime(), eventBody, (short) tenantID });
                 eev.forceDeliveryPriority(DELIVERY_REQUEST_PRIORITY);
               }
           }
